adding constness to function GetEffectCauser and running explicit cast on enum type to be converted to float value in PumpHeartbeat

+++ Plugins/GameFeatures/InteractionSample/Source/InteractionSample/Private/UI/InteractionPresenter.cpp
@@ -77,7 +77,7 @@ void UInteractionPresenter::BP_OnNotificationReceived_PumpHeartbeat(const TInsta
 	                     TEXT("UInteractionPresenter::ViewModel doesn't derive from UInteractionViewModel!")))
 	{
 		const auto* NewHeatbeat = Payload.GetPtr<FAVVMHearbeatPayload>();
-		InteractionViewModel->PumpHeartbeat((NewHeatbeat != nullptr) ? NewHeatbeat->Value : INDEX_NONE);
+		InteractionViewModel->PumpHeartbeat((NewHeatbeat != nullptr) ? NewHeatbeat->Value : static_cast<float>(INDEX_NONE));
 	}
 }
